Move `check_precise_output` and `update_test` to `subtest`
